Moved the speciate function to utils where it belongs
