+++ neat/utils.py
@@ -7,6 +7,7 @@ import torch
 import numpy as np
 
 from neat.phenotype.feed_forward import FeedForwardNet
+from species import Species
 
 logger = logging.getLogger(__name__)
 
@@ -40,7 +41,7 @@ def get_best_genome(population):
 
 
 def create_prediction_map(genomes, dataset, config):
-    '''
+    """
     Creates a prediction mapping of genomes to their 2D arrays of predictions
 
     Each array in the 2D array is the prediction (activation outputs) for the given input
@@ -53,7 +54,7 @@ def create_prediction_map(genomes, dataset, config):
 
     Returns:
         dict of genome to the genome's predictions
-    '''
+    """
     genomes_to_results = {}
     for genome in genomes:
         results = []
@@ -68,17 +69,17 @@ def create_prediction_map(genomes, dataset, config):
 
 
 def random_ensemble_generator(genomes, k=None, limit=None):
-    '''
+    """
     A generator that randomly picks an ensemble from the given genomes of length k
 
     Parameters:
         genomes (iterable): the whole population of genomes to sample from
-        k (None | int): None (for random size ensembles) or the size of ensembles to create 
+        k (None | int): None (for random size ensembles) or the size of ensembles to create
         limit: the number of ensembles to yield before iterable exhaustion
 
     Yields:
        A set of genomes to use in an ensemble
-    '''
+    """
     genomes = list(genomes)
     n = len(genomes)
     seen = set()
@@ -99,19 +100,45 @@ def random_ensemble_generator(genomes, k=None, limit=None):
 
 
 def random_ensemble_generator_for_static_genome(genome, genomes, k=None, limit=None):
-    '''
+    """
     A generator that randomly picks the rest of an ensemble of size k including a given genome
 
     Parameters:
         genome: the genome to include in the resulting ensembles
         genomes (iterable): the whole population of genomes to sample from
-        k (None | int): None (for random size ensembles) or the size of ensembles to create 
+        k (None | int): None (for random size ensembles) or the size of ensembles to create
         limit: the number of ensembles to yield before iterable exhaustion
 
     Yields:
         A set of genomes to use in an ensemble
-    '''
+    """
     genomes = [g for g in genomes if g != genome]
     k = None if k is None else k - 1
     for ensemble in random_ensemble_generator(genomes=genomes, k=k, limit=limit):
         yield {genome, *ensemble}
+
+
+def speciate(genomes, speciation_threshold):
+    """
+    Copied and modified from population.py
+    Creates a list of species, where each species is a list of genomes
+    The speciation takes place based on genomes' genetic diversity
+    """
+
+    species = []
+
+    def speciate(genome):
+        for s in species:
+            if Species.species_distance(genome, s.model_genome) <= speciation_threshold:
+                s.members.append(genome)
+                return
+
+        # Did not match any current species. Create a new one
+        new_species = Species(len(species), genome, 0)
+        new_species.members.append(genome)
+        species.append(new_species)
+
+    for genome in genomes:
+        speciate(genome)
+
+    return [s.members for s in species]
